Move cURLDescription to Pulse
