+++ Sources/Pulse/LoggerStore/LoggerStore+Extensions.swift
@@ -1,0 +1,38 @@
+// The MIT License (MIT)
+//
+// Copyright (c) 2020–2022 Alexander Grebenyuk (github.com/kean).
+
+import Foundation
+
+#if os(iOS) || os(macOS)
+
+extension NetworkTaskEntity {
+    public func cURLDescription() -> String {
+        guard let request = currentRequest ?? originalRequest,
+              let url = request.url else {
+            return "$ curl command generation failed"
+        }
+
+        var components = ["curl -v"]
+
+        components.append("-X \(request.httpMethod ?? "GET")")
+
+        for header in request.headers {
+            let escapedValue = header.value.replacingOccurrences(of: "\"", with: "\\\"")
+            components.append("-H \"\(header.key): \(escapedValue)\"")
+        }
+
+        if let httpBodyData = requestBody?.data {
+            let httpBody = String(decoding: httpBodyData, as: UTF8.self)
+            var escapedBody = httpBody.replacingOccurrences(of: "\\\"", with: "\\\\\"")
+            escapedBody = escapedBody.replacingOccurrences(of: "\"", with: "\\\"")
+            components.append("-d \"\(escapedBody)\"")
+        }
+
+        components.append("\"\(url)\"")
+
+        return components.joined(separator: " \\\n\t")
+    }
+}
+
+#endif
