+++ Sources/PulseUI/Extensions/PulseCore+Extensions.swift
@@ -5,39 +5,6 @@
 import Foundation
 import Pulse
 
-#if os(iOS) || os(macOS)
-
-extension NetworkTaskEntity {
-    func cURLDescription() -> String {
-        guard let request = currentRequest ?? originalRequest,
-              let url = request.url else {
-            return "$ curl command generation failed"
-        }
-
-        var components = ["curl -v"]
-
-        components.append("-X \(request.httpMethod ?? "GET")")
-
-        for header in request.headers {
-            let escapedValue = header.value.replacingOccurrences(of: "\"", with: "\\\"")
-            components.append("-H \"\(header.key): \(escapedValue)\"")
-        }
-
-        if let httpBodyData = requestBody?.data {
-            let httpBody = String(decoding: httpBodyData, as: UTF8.self)
-            var escapedBody = httpBody.replacingOccurrences(of: "\\\"", with: "\\\\\"")
-            escapedBody = escapedBody.replacingOccurrences(of: "\"", with: "\\\"")
-            components.append("-d \"\(escapedBody)\"")
-        }
-
-        components.append("\"\(url)\"")
-
-        return components.joined(separator: " \\\n\t")
-    }
-}
-
-#endif
-
 extension NetworkTaskEntity {
     var requestFileViewerContext: FileViewerViewModel.Context {
         FileViewerViewModel.Context(
@@ -57,9 +24,5 @@ extension NetworkTaskEntity {
             isResponse: true,
             error: decodingError
         )
-    }
-
-    var decodingError: NetworkLogger.DecodingError? {
-        error as? NetworkLogger.DecodingError
     }
 }
